fix(limits): stop leaking connections, respect the configured speed and count traffic
Batch 2 of the review: correctness of the limiter and of the traffic counters.

ConnLimiter leaked the per user tcp connection count and eventually locked
users out ("Limited by ip or conn"):

- A connection rejected by the ip limit had already incremented the count
  and returned without releasing it, so every rejected connection pushed the
  user one step closer to connLimit.
- Only the xray dispatcher ever called DelConnCount. sing counted tcp
  connections and never released them, so ConnLimit behaved as a lifetime
  limit there.
- DelConnCount returned early for non-realtime nodes, so even xray never
  decremented on those nodes.

The increment/decrement now live in addConnCount/delConnCount helpers, the
rejection path rolls its own increment back, the counter is released in both
modes and sing releases it from the tracked connection close callback.
hysteria2 is not affected: its client addresses are UDP, so CheckLimit never
accounted a tcp connection for it.

common/rate: Conn.Read/Write waited for len(b) instead of the bytes actually
transferred, charging every read a whole buffer worth of tokens. With a
32KiB buffer a one byte read consumed 32KiB and the connection was throttled
to a fraction of the configured speed.

common/counter: ConnCounter.Read/Write used Store, so each operation threw
away everything counted before it and a long lived connection only reported
its last chunk. They now use Add, like ReadBuffer/WriteBuffer already did.

common/counter: TrafficCounter.Rx/Tx took an int, truncating the cumulative
stream counters above 2GiB on the 32bit builds the project ships (386,
armv7, mips). They take an int64 now.

+++ common/counter/traffic.go
@@ -63,12 +63,15 @@ func (c *TrafficCounter) Delete(id string) {
 	c.counters.Delete(id)
 }
 
-func (c *TrafficCounter) Rx(id string, n int) {
+// Rx and Tx take an int64: they receive the byte counts of a whole stream and
+// an int truncates them to 32 bits on the 32 bit targets this project builds
+// (386, armv7, mips), silently losing traffic above 2GiB per stream.
+func (c *TrafficCounter) Rx(id string, n int64) {
 	cts := c.GetCounter(id)
-	cts.DownCounter.Add(int64(n))
+	cts.DownCounter.Add(n)
 }
 
-func (c *TrafficCounter) Tx(id string, n int) {
+func (c *TrafficCounter) Tx(id string, n int64) {
 	cts := c.GetCounter(id)
-	cts.UpCounter.Add(int64(n))
+	cts.UpCounter.Add(n)
 }
